team_hypeshot/requirements.txt
carvekit==4.1.0
IPTCInfo3==2.1.4
Pillow
torch==1.11.0
